backend/requirements.txt
google-adk>=1.0.0
google-genai
fastapi
uvicorn[standard]
websockets
httpx
PyMuPDF
chromadb>=0.6.0
aiosqlite>=0.20.0
